+++ multibranch/src/main/java/org/jenkinsci/plugins/workflow/multibranch/BranchJobProperty.java
@@ -39,9 +39,13 @@ import org.jenkinsci.plugins.workflow.job.WorkflowJobProperty;
  */
 public class BranchJobProperty extends WorkflowJobProperty {
 
-    private /*@Nonnull once initialized */ Branch branch;
+    private @Nonnull Branch branch;
 
-    public synchronized Branch getBranch() {
+    BranchJobProperty(@Nonnull Branch branch) {
+        this.branch = branch;
+    }
+
+    public synchronized @Nonnull Branch getBranch() {
         return branch;
     }
 
@@ -65,6 +69,8 @@ public class BranchJobProperty extends WorkflowJobProperty {
             }
         };
     }
+
+    // TODO make WorkflowJob.isBuildable false if !branch.isBuildable to handle orphaned projects
 
     @Extension public static class DescriptorImpl extends JobPropertyDescriptor {